Initial commit
This is the initial commit with some base pieces working for
interacting with a command line client to do some basic checks
for whether Juju is installed, Microk8s is installed, etc and
some basic framework for running commands.

The checks for microk8s/juju should be moved to dependencies
defined in the snapcraft.yaml file but require that both of the
other snaps are delivered via strict confinement (which is not
the case today).