chore: Format code and ensure tox targets work
Format the code using the black python code formatter. This may be a bit
controversial since it has some defaults outside what is standard in
OpenStack projects, but it can enforce consistent formatting and that's
a bit more practical.

This also introduces a couple of minimal unit tests and some coverage
bits to ensure that everything runs a-ok.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>