Refactor Jujuhelper class
Juju connection is lost when multiple actions
are fired one after another. In addition, there
are exceptions on the console that are not caught
properly. Refactored the JujuHelper class to
connect to controller once per command instead of
each connect/disconnect for each step in the command.

Note: JujuHelper should be replaced with zaza library
once zaza supports libjuju 3.0